Documentation for QECC.

QECC.Bitflip3

QECC.CSS

QECC.Cleve8

QECC.Perfect5

QECC.Phaseflip3

QECC.RepCode

QECC.Shor9

QECC.Steane7

QECC.Surface

QECC.Toric

QECC.code_n

QECC.code_s

QECC.distance

QECC.hgp

QECC.parity_matrix

QECC.parity_matrix_x

QECC.parity_matrix_z

QECC.Bitflip3 — Type

Three-qubit bit-flip code.

Bitflip3 is a quantum error correction code that corrects single-qubit bit-flip error and does not detect any phase-flip errors.

QECC.CSS — Type

CSS(Hx, Hz)

An arbitrary CSS error correcting code defined by its X and Z checks. Hx and Hz are the parity check matrices of the X and Z stabilizers, respectively.

QECC.Cleve8 — Type

A pedagogical example of a quantum error correcting [8,3] code used in (Cleve and Gottesman, 1997).

QECC.Perfect5 — Type

Perfect five-qubit code (Lafiamme et al., 1996).

Perfect5 code is the smallest qubit stabilizer code to correct a single-qubit error.

QECC.Phaseflip3 — Type

Three-qubit phase-flip code.

Phaseflip3 is a quantum error correction code that corrects single-qubit phase-flip error and does not detect any bit-flip errors.

QECC.RepCode — Type

Repetition code.

Repetition code is a classical error correction code that repeats the same bit n times.

QECC.Shor9 — Type

Shor9 code (Shor, 1995).

Nine-qubit CSS code that is the first quantum error-correcting code.

QECC.Steane7 — Type

Steane code (Steane, 1996).

QECC.Surface — Type

The planar surface code refers to the code (Kitaev, 2003) in a 2D lattice with open boundaries.

Illustration of a 3×2 surface code, where qubits are located on the edges:

|---1--(Z)--2---|---3---|
|  (X)  7       8       o
|---4---|---5---|---6---|
|       o       o       o
|       |       |       |

The surface code has open boundary conditions, unlike the toric code. To this end, we remove qubits (denoted by "o") and parity checks on the right and bottom sides.

More information can be seen in (Fowler et al., 2012).

QECC.Toric — Type

The Toric code (Kitaev, 2003).

Illustration of a 2x2 toric code, where qubits are located on the edges:

|--1-(Z)-2--|
| (X) 5     6
|--3--|--4--|
|     7     8
|     |     |

It is important to note that the toric code has periodic boundary conditions, which means that the top and bottom sides are essentially glued together, as are the left and right sides.

QECC.code_n — Method

The number of physical qubits in a code.

QECC.code_s — Method

The number of stabilizer checks in a code. They might not be all linearly independent, thus code_s >= code_n-code_k. For the number of linearly independent checks you can use LinearAlgebra.rank.

QECC.distance — Function

The distance of a code.

QECC.hgp — Method

Hypergraph product of two classical codes.

QECC.parity_matrix — Function

The parity check matrix of a code.

QECC.parity_matrix_x — Function

Parity check boolean matrix of a code (only the X entries in the tableau, i.e. the checks for Z errors). Only CSS codes have this method.

QECC.parity_matrix_z — Function

Parity check boolean matrix of a code (only the Z entries in the tableau, i.e. the checks for X errors). Only CSS codes have this method.