Documentation for QECC.

QECC.AbstractCECC

QECC.AbstractCSSCode

QECC.AbstractECC

QECC.AbstractQECC

QECC.Bitflip3

QECC.CSS

QECC.Cleve8

QECC.Perfect5

QECC.Phaseflip3

QECC.RepCode

QECC.Shor9

QECC.Steane7

QECC.Surface

QECC.Toric

QECC.code_k

QECC.code_n

QECC.code_s

QECC.distance

QECC.hgp

QECC.parity_matrix

QECC.parity_matrix_x

QECC.parity_matrix_z

QECC.AbstractCECC — Type

AbstractCECC <: AbstractECC

Abstract type for classical error correction code.

QECC.AbstractCSSCode — Type

AbstractCSSCode <: AbstractQECC

Abstract type for Calderbank-Shor-Steane (CSS) code.

QECC.AbstractECC — Type

AbstractECC

Abstract type for error correction code.

QECC.AbstractQECC — Type

AbstractQECC <: AbstractECC

Abstract type for quantum error correction code.

QECC.Bitflip3 — Type

Bitflip3 <: AbstractQECC

Bitflip3 is a three-qubit bit-flip code that corrects single-qubit bit-flip error and does not detect any phase-flip errors.

QECC.CSS — Type

CSS <: AbstractCSSCode
CSS(Hx, Hz)

An arbitrary CSS error correction code defined by its X and Z checks.

Fields

Hx: The parity check matrix of the X stabilizers.

Hz: The parity check matrix of the Z stabilizers.

QECC.Cleve8 — Type

Cleve8 <: AbstractQECC

A pedagogical example of a quantum error correcting [8,3] code used in (Cleve and Gottesman, 1997).

QECC.Perfect5 — Type

Perfect5 <: AbstractQECC

Perfect5 code (Lafiamme et al., 1996) is the smallest qubit stabilizer code to correct a single-qubit error.

QECC.Phaseflip3 — Type

Phaseflip3 <: AbstractQECC

Phaseflip3 is a three-qubit phase-flip code that corrects single-qubit phase-flip error and does not detect any bit-flip errors.

QECC.RepCode — Type

RepCode <: AbstractCECC
RepCode(n)

Repetition code is a classical error correction code that repeats the same bit n times.

Fields

n: The number of times to repeat the bit.

QECC.Shor9 — Type

Shor9 <: AbstractCSSCode

Shor9 code (Shor, 1995)is a nine-qubit CSS code that is the first quantum error-correcting code.

QECC.Steane7 — Type

Steane7 <: AbstractCSSCode

Steane code (Steane, 1996).

QECC.Surface — Type

Surface <: AbstractCSSCode
Surface(dx, dz)

The planar surface code refers to the code (Kitaev, 2003) in a 2D lattice with open boundaries. Illustration of a 3×2 surface code, where qubits are located on the edges:

|---1--(Z)--2---|---3---|
|  (X)  7       8       o
|---4---|---5---|---6---|
|       o       o       o
|       |       |       |

The surface code has open boundary conditions, unlike the toric code. To this end, we remove qubits (denoted by "o") and parity checks on the right and bottom sides.

More information can be seen in (Fowler et al., 2012).

Fields

dx: The number of qubits in the x direction.

dz: The number of qubits in the z direction.

QECC.Toric — Type

Toric <: AbstractCSSCode
Toric(dx, dz)

The Toric code (Kitaev, 2003). Illustration of a 2x2 toric code, where qubits are located on the edges:

|--1-(Z)-2--|
| (X) 5     6
|--3--|--4--|
|     7     8
|     |     |

It is important to note that the toric code has periodic boundary conditions, which means that the top and bottom sides are essentially glued together, as are the left and right sides.

Fields

dx: The number of qubits in the x direction.

dz: The number of qubits in the z direction.

QECC.code_k — Function

code_k(c::AbstractECC)

The number of logical qubits in a error correction code.

QECC.code_n — Method

code_n(c::AbstractECC)

The number of physical qubits in a error correction code.

QECC.code_s — Method

code_s(c::AbstractECC)

The number of stabilizers in a error correction code. They might not be all linearly independent, thus code_s >= code_n-code_k. For the number of linearly independent checks you can use LinearAlgebra.rank.

QECC.distance — Function

distance(c::AbstractECC)

The code distance of a error correction code.

QECC.hgp — Method

hgp(h₁::AbstractMatrix{Bool}, h₂::AbstractMatrix{Bool})

Hypergraph product of two classical codes.

QECC.parity_matrix — Function

parity_matrix(c::AbstractECC)

The parity check matrix of a error correction code.

QECC.parity_matrix_x — Function

parity_matrix_x(c::AbstractCSSCode)

Parity check boolean matrix of a code (only the X entries in the tableau, i.e. the checks for Z errors). Only CSS codes have this method.

QECC.parity_matrix_z — Function

parity_matrix_z(c::AbstractCSSCode)

Parity check boolean matrix of a code (only the Z entries in the tableau, i.e. the checks for X errors). Only CSS codes have this method.